requirements.txt
# Core Earth Engine and Geospatial packages
earthengine-api>=1.6.3
folium>=0.20.0
geopandas>=1.0.1
shapely>=2.0.7

# Data Science packages
pandas>=2.3.0
numpy>=2.0.0
matplotlib>=3.9.0
seaborn>=0.13.0
scipy>=1.13.0
scikit-learn>=1.6.0

# Web App packages
streamlit>=1.48.0
streamlit-folium>=0.15.0
leafmap>=0.28.0

# Utility packages
requests>=2.32.0
pillow>=11.0.0
pyyaml>=6.0.0
geojson>=3.0.0

# Jupyter packages (optional - for notebook usage)
jupyter>=1.1.0
jupyterlab>=4.4.0
ipywidgets>=8.1.0

# Google Cloud packages (for Earth Engine)
google-api-python-client>=2.179.0
google-auth>=2.40.0
google-cloud-storage>=3.3.0
